Move https tests for OpenRosaInterface into client specific tests
It feels like the behaviour (although the intention is the same)
is different enough to warrant the tests being implementation
specific
