+++ collect_app/src/test/java/org/odk/collect/android/http/HttpClientConnectionTest.java
@@ -1,22 +1,43 @@
 package org.odk.collect.android.http;
 
+import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.robolectric.RobolectricTestRunner;
 
 import java.io.IOException;
 
+import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
+import okhttp3.mockwebserver.RecordedRequest;
+import okhttp3.tls.internal.TlsUtil;
+
+import static org.hamcrest.Matchers.equalTo;
+import static org.junit.Assert.assertThat;
 
 @RunWith(RobolectricTestRunner.class)
 public class HttpClientConnectionTest extends OpenRosaHttpInterfaceTest {
 
     @Override
-    protected void startHttpsMockWebServer(MockWebServer mockWebServer) throws IOException {
-        mockWebServer.start(8443);
-    }
-
-    @Override
     protected OpenRosaHttpInterface buildSubject() {
         return new HttpClientConnection();
+    }
+
+    @Test
+    public void executeGetRequest_withCredentials_whenHttpsPort_retriesWithCredentials() throws Exception  {
+        MockWebServer httpsPortMockWebServer = new MockWebServer();
+        httpsPortMockWebServer.start(8443);
+
+        httpsPortMockWebServer.enqueue(new MockResponse()
+                .setResponseCode(401)
+                .addHeader("WWW-Authenticate: Basic realm=\"protected area\"")
+                .setBody("Please authenticate."));
+        httpsPortMockWebServer.enqueue(new MockResponse());
+
+        buildSubject().executeGetRequest(httpsPortMockWebServer.url("").uri(), null, new HttpCredentials("user", "pass"));
+
+        assertThat(httpsPortMockWebServer.getRequestCount(), equalTo(2));
+        httpsPortMockWebServer.takeRequest();
+        RecordedRequest request = httpsPortMockWebServer.takeRequest();
+        assertThat(request.getHeader("Authorization"), equalTo("Basic dXNlcjpwYXNz"));
     }
 }
